adapting api to accept input from dnd api ip
