+++ src/intTest/presentation/HomeController.java
@@ -1,4 +1,4 @@
-package org.jointheleague.api.giraffe.Giraffe.Search.presentation;
+package org.jointheleague.level7.giraffe.presentation;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Controller;
