--- src/main/java/org/jointheleague/api/giraffe/Giraffe/Search/Output.java
@@ -1,5 +1,0 @@
-package org.jointheleague.api.giraffe.Giraffe.Search;
-
-public class Output {
-
-}
