+++ src/main/java/org/jointheleague/level7/giraffe/Output.java
@@ -1,0 +1,5 @@
+package org.jointheleague.level7.giraffe;
+
+public class Output {
+
+}
