+++ src/main/java/org/jointheleague/level7/giraffe/presentation/HomeController.java
@@ -1,11 +1,13 @@
-package org.jointheleague.api.giraffe.Giraffe.Search.presentation;
+package org.jointheleague.level7.giraffe.presentation;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ResponseStatus;
 
 @Controller
+@CrossOrigin
 public class HomeController {
 
     @GetMapping("/")
