+++ src/main/java/org/jointheleague/level7/giraffe/presentation/LocController.java
@@ -1,20 +1,21 @@
-package org.jointheleague.api.giraffe.Giraffe.Search.presentation;
+package org.jointheleague.level7.giraffe.presentation;
 
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiResponse;
 import io.swagger.annotations.ApiResponses;
-import org.jointheleague.api.giraffe.Giraffe.Search.repository.dto.Result;
-import org.jointheleague.api.giraffe.Giraffe.Search.service.LocService;
+import org.jointheleague.level7.giraffe.repository.dto.Result;
+import org.jointheleague.level7.giraffe.service.LocService;
 import org.springframework.http.HttpStatus;
 import org.springframework.util.CollectionUtils;
+import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
-
 @RestController
+@CrossOrigin
 public class LocController {
 
     private final LocService locService;
@@ -26,8 +27,7 @@ public class LocController {
     @GetMapping("/searchLocResults")
     @ApiOperation(value = "Searches for articles matching the search term",
             notes = "Response may include multiple Result values.",
-            response = Result.class,
-            responseContainer="List")
+            response = Result.class)
     @ApiResponses(value = {
             @ApiResponse(code = 200, message = "Result(s) found"),
             @ApiResponse(code = 404, message = "Result(s) not found")
