+++ src/main/java/org/jointheleague/level7/giraffe/repository/dto/LocResponse.java
@@ -1,14 +1,9 @@
-package org.jointheleague.api.giraffe.Giraffe.Search.repository.dto;
+package org.jointheleague.level7.giraffe.repository.dto;
 
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 import java.util.List;
-
-
-import java.util.List;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
 
 
 public class LocResponse {
